backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
pytest==8.1.1
pytest-flask==1.3.0
requests==2.31.0
openai==1.30.1
psycopg2-binary==2.9.9
